#41 Moved interfaces to services folder
