+++ src/app-shared/core/core.module.ts
@@ -11,18 +11,18 @@ import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { ModuleWithProviders, NgModule, Provider, Type } from '@angular/core';
 import { FormControlErrorsComponent } from './components/form-control-errors/form-control-errors.component';
-import { IAccountService } from './interface/account.interface';
-import { IApiEndpointBuilderService } from './interface/api-endpoint-builder.interface';
-import { INotificationService } from './interface/notification.interface';
-import { IStringTemplateService } from './interface/string-template.interface';
-import { ITranslationService } from './interface/translation.interface';
-import { IWebRequestService } from './interface/web-request.interface';
 import { TranslatePipe } from './pipes/translate.pipe';
+import { IAccountService } from './service/account.interface';
 import { AccountService } from './service/account.service';
+import { IApiEndpointBuilderService } from './service/api-endpoint-builder.interface';
 import { ApiEndpointBuilderService } from './service/api-endpoint-builder.service';
+import { INotificationService } from './service/notification.interface';
 import { NotificationService } from './service/notification.service';
+import { IStringTemplateService } from './service/string-template.interface';
 import { StringTemplateService } from './service/string-template.service';
+import { ITranslationService } from './service/translation.interface';
 import { TranslationService } from './service/translation.service';
+import { IWebRequestService } from './service/web-request.interface';
 import { WebRequestService } from './service/web-request.service';
 
 const SHARED_COMPONENTS: Type<any>[] = [
