--- src/app-shared/core/interface/account-credentials.ts
@@ -1,14 +1,0 @@
-/**
- * @license
- * Copyright (c) 2018 THREEANGLE SOFTWARE SOLUTIONS SRL
- * Available under MIT license webFrame/LICENSE
- */
-
-/**
- * Provides IAccountCredentials.
- */
-
-export interface IAccountCredentials {
-  username: string;
-  password: string;
-}
