+++ src/app-shared/core/service/account.service.ts
@@ -9,11 +9,9 @@ import { Observable, of } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { environment } from '../../../environments/environment';
 import { Empty } from '../data/empty.dm';
-import { IAccountCredentials } from '../interface/account-credentials';
-import { IAccountInformation } from '../interface/account-information';
-import { IAccountService } from '../interface/account.interface';
-import { ServerApi } from '../interface/api-endpoint-builder.interface';
-import { IWebRequestService, RequestContentType } from '../interface/web-request.interface';
+import { IAccountCredentials, IAccountInformation, IAccountService } from './account.interface';
+import { ServerApi } from './api-endpoint-builder.interface';
+import { IWebRequestService, RequestContentType } from './web-request.interface';
 
 @Injectable()
 export class AccountService implements IAccountService {
