+++ src/app-shared/core/service/json-converter.interface.ts
