+++ src/app-shared/core/service/notification.service.spec.ts
@@ -6,8 +6,8 @@
 
 import { inject, TestBed } from '@angular/core/testing';
 import { createTranslationServiceStub } from 'app-shared/test-utils';
-import { ITranslationService } from '../interface/translation.interface';
 import { NotificationService } from './notification.service';
+import { ITranslationService } from './translation.interface';
 
 describe('NotificationService', () => {
   beforeEach(() => {
