+++ src/app-shared/core/service/notification.service.ts
@@ -5,8 +5,8 @@
  */
 
 import { Inject, Injectable } from '@angular/core';
-import { INotificationConfig, INotificationService } from '../interface/notification.interface';
-import { ITranslationService } from '../interface/translation.interface';
+import { INotificationConfig, INotificationService } from './notification.interface';
+import { ITranslationService } from './translation.interface';
 
 @Injectable()
 export class NotificationService implements INotificationService {
