+++ src/app-shared/core/service/string-template.interface.ts
