+++ src/app-shared/core/service/translation.interface.ts
