+++ src/app-shared/core/service/translation.service.ts
@@ -5,7 +5,7 @@
  */
 
 import { Injectable } from '@angular/core';
-import { ITranslationService, TranslationPhrase, TranslationPhraseArgs } from '../interface/translation.interface';
+import { ITranslationService, TranslationPhrase, TranslationPhraseArgs } from './translation.interface';
 
 @Injectable()
 export class TranslationService implements ITranslationService {
