+++ src/app-shared/core/service/web-request.interface.ts
@@ -7,8 +7,8 @@
 import { InjectionToken } from '@angular/core';
 import { Observable } from 'rxjs';
 
+import { Dictionary } from '../interface/dictionary';
 import { ServerApi } from './api-endpoint-builder.interface';
-import { Dictionary } from './dictionary';
 
 /**
  * Custom type for the URL parameters' values
