+++ src/app-shared/core/service/web-request.service.ts
@@ -7,8 +7,8 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Inject, Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { IApiEndpointBuilderService } from '../interface/api-endpoint-builder.interface';
-import { IWebRequestService, RequestConfiguration, RequestContentType } from '../interface/web-request.interface';
+import { IApiEndpointBuilderService } from './api-endpoint-builder.interface';
+import { IWebRequestService, RequestConfiguration, RequestContentType } from './web-request.interface';
 
 export type RequestStrategy<T> = (url: string, headers: HttpHeaders, body?: string) => Observable<T>;
